v0.7.0
A lot of progress has been made since v0.6.0 and It feels appropriate to bump the version to v0.7.0.

Milestones in v0.6.x:
  - Redesigned s16's ISA (along with t16's for backwards compatability)
  - Updated and refactored s16' assembler
  - Accessing the instruction cache and instruction prefetch
  - Work started on the main pipeline loop 
 
Heavy brainstorming and planning has begun for the pipeline and the required methods. Along with,  figuring out the complexity/direction I want for s16's pipeline. 
v0.7.x will focus on completing the planning of the pipeline, creating any methods required, and getting instructions to modify s16's state(finally >_>)