v0.7.0-1
Progress will be slow in v0.7.x as this'll be a version aimed to flesh out how the pipeline will work. All instructions are ready to be implemented and without concrete plan for the pipeline, there'd be a need for a huge refactor for any changes.

This document (likely many others too) will go over the pipeline and fully flesh out the plan for s16's pipeline.